Reorg examples namespace
